we need to be able to configure if pace is running behind a proxy
problem is, that we want to set the secure flag of
the cookies. but since the connection between pace and the
proxy is not encrypted, express does not set any cookie.
we need to be able to tell express that the connection is
encrypted but only beteen the proxy and the user.