Absolute edit URI.
On the success page, in the email and in list of all participants.
First I discovered that the previous solution: req.header('origin') always
returned 'undefined' in Firefox - God knows why... chrome and safari did not
have the problem. Nevertheless I than saw, that we did not use absolute URIs
in the emails yet and therefore the people would not be able to navigate out
of the email to the page, so I fixed it by introducing 'pace-url' as part of our
config. It can be also provided as env PACE_URL. @cz8s please update the setup
on our prod env.

@mduemcke I hope, I did not implemented something you already have laying
around in your fork :)